version 1.0.0

+++ pages/api/v1/deleteGroup.js
@@ -4,24 +4,26 @@ import Users from "../../../models/userModel";
 
 export default async function deleteGroup(req, res) {
   if (req.method === "DELETE") {
-    try {
-      await connectMongo();
-      console.log("CONNECTED TO MONGO");
-
-      const groupId = req.body.id;
-      const group = await Groups.findByIdAndDelete(groupId);
-      if (group) {
-        const userId = group.createdBy;
-        const user = await Users.findById(userId);
-        let createdGroups = user.createdGroups;
-        let index = createdGroups.indexOf(groupId);
-        user.createdGroups.splice(index, 1);
-        user.save();
-        res.json({ group, user });
+    if (req.query.apiSecret === process.env.API_SECRET) {
+      try {
+        await connectMongo();
+        const groupId = req.body.id;
+        const group = await Groups.findByIdAndDelete(groupId);
+        if (group) {
+          const userId = group.createdBy;
+          const user = await Users.findById(userId);
+          let createdGroups = user.createdGroups;
+          let index = createdGroups.indexOf(groupId);
+          user.createdGroups.splice(index, 1);
+          user.save();
+          res.json({ group, user });
+        }
+      } catch (error) {
+        console.log(error);
+        res.json({ error });
       }
-    } catch (error) {
-      console.log(error);
-      res.json({ error });
+    } else {
+      res.json("Not Authorized to make API Request");
     }
   } else {
     res.json("Wrong Method");
